Move implementation into OsuPlayfield
